--- a/distance-moved.cpp
+++ b/distance-moved.cpp
@@ -60,6 +60,8 @@ float robot_angle;    // degrees
 float robot_velocity; // mm/s
 float robot_omega;    // deg/s
 
+
+
 /***
  * Local variables
  *
@@ -121,12 +123,10 @@ void update_encoders()
     encoder_right_total += right_count;
     encoder_left_total += left_count;
 
-    int encoder_sum = right_count + left_count;
-    robot_velocity = LOOP_FREQUENCY * MM_PER_COUNT * encoder_sum;
-    robot_distance = MM_PER_COUNT * (encoder_right_total + encoder_left_total);
+    robot_velocity = LOOP_FREQUENCY * MM_PER_COUNT * (right_count + left_count);
+    robot_omega = LOOP_FREQUENCY * DEG_PER_COUNT * (right_count - left_count);
 
-    int encoder_diff = right_count - left_count;
-    robot_omega = LOOP_FREQUENCY * DEG_PER_COUNT * encoder_diff;
+    robot_distance = MM_PER_COUNT * (encoder_right_total + encoder_left_total);
     robot_angle = DEG_PER_COUNT * (encoder_right_total - encoder_left_total);
 }