--- a/distance-moved.cpp
+++ b/distance-moved.cpp
@@ -122,9 +122,13 @@ void update_encoders()
 
     encoder_right_total += right_count;
     encoder_left_total += left_count;
+    // calculate and low pass filter the speeds
+    //TODO: how does the filtering affect the controller?
+    float velocity = LOOP_FREQUENCY * MM_PER_COUNT * (right_count + left_count);
+    robot_velocity += 0.5*(velocity - robot_velocity);
 
-    robot_velocity = LOOP_FREQUENCY * MM_PER_COUNT * (right_count + left_count);
-    robot_omega = LOOP_FREQUENCY * DEG_PER_COUNT * (right_count - left_count);
+    float omega = LOOP_FREQUENCY * DEG_PER_COUNT * (right_count - left_count);
+    robot_omega += 0.5*(omega - robot_omega);
 
     robot_distance = MM_PER_COUNT * (encoder_right_total + encoder_left_total);
     robot_angle = DEG_PER_COUNT * (encoder_right_total - encoder_left_total);