--- a/distance-moved.cpp
+++ b/distance-moved.cpp
@@ -60,8 +60,6 @@ float robot_angle;    // degrees
 float robot_velocity; // mm/s
 float robot_omega;    // deg/s
 
-
-
 /***
  * Local variables
  *
@@ -125,10 +123,10 @@ void update_encoders()
     // calculate and low pass filter the speeds
     //TODO: how does the filtering affect the controller?
     float velocity = LOOP_FREQUENCY * MM_PER_COUNT * (right_count + left_count);
-    robot_velocity += 0.5*(velocity - robot_velocity);
+    robot_velocity += 0.5 * (velocity - robot_velocity);
 
     float omega = LOOP_FREQUENCY * DEG_PER_COUNT * (right_count - left_count);
-    robot_omega += 0.5*(omega - robot_omega);
+    robot_omega += 0.5 * (omega - robot_omega);
 
     robot_distance = MM_PER_COUNT * (encoder_right_total + encoder_left_total);
     robot_angle = DEG_PER_COUNT * (encoder_right_total - encoder_left_total);