--- a/interpreter.cpp
+++ b/interpreter.cpp
@@ -34,6 +34,7 @@
   SOFTWARE.
 */
 #include "digitalWriteFast.h"
+#include "tests.h"
 #include "public.h"
 #include <Arduino.h>
 #include <EEPROM.h>
@@ -1142,7 +1143,7 @@ const PROGMEM fptr PROGMEM cmd2[] =
         not_implemented,                    // 'n'
         not_implemented,                    // 'o'
         not_implemented,                    // 'p'
-        not_implemented,                    // 'q'
+        cmd_test_runner,                    // 'q'
         print_encoder_setup,                // 'r'
         print_switches,                     // 's'
         not_implemented,                    // 't'