--- a/motor_control.cpp
+++ b/motor_control.cpp
@@ -33,13 +33,21 @@
 */
 #include "digitalWriteFast.h"
 #include "hardware_pins.h"
+#include "pid_v1.h"
 #include "public.h"
 #include "robot_config.h"
 #include <Arduino.h>
-
 /***
  * Global variables
  */
+
+float fwd_kp = 0.003;
+float fwd_ki = 0.001;
+float fwd_kd = 0.000;
+float fwd_set_speed;
+float fwd_output;
+
+PID fwd_controller(&robot_velocity, &fwd_output, &fwd_set_speed, fwd_kp, fwd_ki, fwd_kd, DIRECT);
 
 void motorSetup()
 {