--- a/motor_control.cpp
+++ b/motor_control.cpp
@@ -42,41 +42,49 @@
  */
 
 float fwd_kp = 0.010;
-float fwd_ki = 1.00;
+float fwd_ki = 0.5;
 float fwd_kd = 0.000;
 float fwd_set_speed;
 float fwd_volts;
 
-
 float rot_kp = 0.010;
-float rot_ki = 1.00;
+float rot_ki = 0.5;
 float rot_kd = 0.000;
 float rot_set_speed;
 float rot_volts;
 
+bool flag_controllers_use_ff = true;
+
 PID fwd_controller(&robot_velocity, &fwd_volts, &fwd_set_speed, fwd_kp, fwd_ki, fwd_kd);
 PID rot_controller(&robot_omega, &rot_volts, &rot_set_speed, rot_kp, rot_ki, rot_kd);
 
-enum { PWM_488_HZ, PWM_3900_HZ, PWM_31_KHZ };
-void pwmSetup(int frequency = PWM_488_HZ) {
-  switch (frequency) {
+enum
+{
+    PWM_488_HZ,
+    PWM_3900_HZ,
+    PWM_31_KHZ
+};
+void pwmSetup(int frequency = PWM_488_HZ)
+{
+    switch (frequency)
+    {
     case PWM_31_KHZ:
-      // Divide by 1. frequency = 31.25 kHz;
-      bitClear(TCCR1B, CS11);
-      bitSet(TCCR1B, CS10);
-      break;
+        // Divide by 1. frequency = 31.25 kHz;
+        bitClear(TCCR1B, CS11);
+        bitSet(TCCR1B, CS10);
+        break;
     case PWM_3900_HZ:
-      // Divide by 8. frequency = 3.91 kHz;
-      bitSet(TCCR1B, CS11);
-      bitClear(TCCR1B, CS10);
-      break;
+        // Divide by 8. frequency = 3.91 kHz;
+        bitSet(TCCR1B, CS11);
+        bitClear(TCCR1B, CS10);
+        break;
     case PWM_488_HZ:
     default:
-      // Divide by 64. frequency = 488Hz;
-      bitSet(TCCR1B, CS11);
-      bitSet(TCCR1B, CS10);
-      break;
-  }
+        // Divide by 64. frequency = 488Hz;
+        bitSet(TCCR1B, CS11);
+        bitSet(TCCR1B, CS10);
+        break;
+    }
 }
 
 void motorSetup()
@@ -98,6 +106,7 @@ void motorSetup()
     rot_controller.SetMode(AUTOMATIC); // turns on the controller. Set to manual for off.
 }
 
+//TODO: if controllers can be disabled, we can apply fixed voltage/pwm to motors
 void update_motors()
 {
     rot_controller.Compute();
@@ -115,15 +124,18 @@ void update_motors()
     left_volts -= rot_volts;
     right_volts += rot_volts;
 
-    float fwd_ff = fwd_set_speed * k_velocity_ff;
-    float rot_ff = rot_set_speed * (WHEEL_SEPARATION / (2 * 57.29)) * k_velocity_ff;
-    // rot_ff = 0;
+    if (flag_controllers_use_ff)
+    {
+        float fwd_ff = fwd_set_speed * k_velocity_ff;
+        float rot_ff = rot_set_speed * (WHEEL_SEPARATION / (2 * 57.29)) * k_velocity_ff;
+        // rot_ff = 0;
 
-    left_volts += fwd_ff;
-    right_volts += fwd_ff;
+        left_volts += fwd_ff;
+        right_volts += fwd_ff;
 
-    left_volts -= rot_ff;
-    right_volts += rot_ff;
+        left_volts -= rot_ff;
+        right_volts += rot_ff;
+    }
 
     setMotorVolts(left_volts, right_volts);
     // setMotorVolts(1.4, 1.4);