--- a/pid_v1.cpp
+++ b/pid_v1.cpp
@@ -30,7 +30,7 @@ PID::PID(float *Input, float *Output, float *Setpoint,
     PID::SetOutputLimits(0, 255); //default output limit corresponds to
                                   //the arduino pwm limits
 
-    SampleTime = 10; //default Controller Sample Time is 0.1 seconds
+    SampleTime = 10; //default Controller Sample Time is 0.01 seconds
 
     PID::SetControllerDirection(ControllerDirection);
     PID::SetTunings(Kp, Ki, Kd, POn);