--- a/pid_v1.cpp
+++ b/pid_v1.cpp
@@ -19,8 +19,8 @@
  *    The parameters specified here are those for for which we can't set up
  *    reliable defaults, so we need to have the user set them.
  ***************************************************************************/
-PID::PID(double *Input, double *Output, double *Setpoint,
-         double Kp, double Ki, double Kd, int POn, int ControllerDirection)
+PID::PID(float *Input, float *Output, float *Setpoint,
+         float Kp, float Ki, float Kd, int POn, int ControllerDirection)
 {
     myOutput = Output;
     myInput = Input;
@@ -43,8 +43,8 @@ PID::PID(double *Input, double *Output, double *Setpoint,
  *    to use Proportional on Error without explicitly saying so
  ***************************************************************************/
 
-PID::PID(double *Input, double *Output, double *Setpoint,
-         double Kp, double Ki, double Kd, int ControllerDirection)
+PID::PID(float *Input, float *Output, float *Setpoint,
+         float Kp, float Ki, float Kd, int ControllerDirection)
     : PID::PID(Input, Output, Setpoint, Kp, Ki, Kd, P_ON_E, ControllerDirection)
 {
 }
@@ -64,9 +64,9 @@ bool PID::Compute()
     if (timeChange >= SampleTime)
     {
         /*Compute all the working error variables*/
-        double input = *myInput;
-        double error = *mySetpoint - input;
-        double dInput = (input - lastInput);
+        float input = *myInput;
+        float error = *mySetpoint - input;
+        float dInput = (input - lastInput);
         outputSum += (ki * error);
 
         /*Add Proportional on Measurement, if P_ON_M is specified*/
@@ -79,7 +79,7 @@ bool PID::Compute()
             outputSum = outMin;
 
         /*Add Proportional on Error, if P_ON_E is specified*/
-        double output;
+        float output;
         if (pOnE)
             output = kp * error;
         else
@@ -108,7 +108,7 @@ bool PID::Compute()
  * it's called automatically from the constructor, but tunings can also
  * be adjusted on the fly during normal operation
  ******************************************************************************/
-void PID::SetTunings(double Kp, double Ki, double Kd, int POn)
+void PID::SetTunings(float Kp, float Ki, float Kd, int POn)
 {
     if (Kp < 0 || Ki < 0 || Kd < 0)
         return;
@@ -120,7 +120,7 @@ void PID::SetTunings(double Kp, double Ki, double Kd, int POn)
     dispKi = Ki;
     dispKd = Kd;
 
-    double SampleTimeInSec = ((double)SampleTime) / 1000;
+    float SampleTimeInSec = ((float)SampleTime) / 1000;
     kp = Kp;
     ki = Ki * SampleTimeInSec;
     kd = Kd / SampleTimeInSec;
@@ -136,7 +136,7 @@ void PID::SetTunings(double Kp, double Ki, double Kd, int POn)
 /* SetTunings(...)*************************************************************
  * Set Tunings using the last-rembered POn setting
  ******************************************************************************/
-void PID::SetTunings(double Kp, double Ki, double Kd)
+void PID::SetTunings(float Kp, float Ki, float Kd)
 {
     SetTunings(Kp, Ki, Kd, pOn);
 }
@@ -148,7 +148,7 @@ void PID::SetSampleTime(int NewSampleTime)
 {
     if (NewSampleTime > 0)
     {
-        double ratio = (double)NewSampleTime / (double)SampleTime;
+        float ratio = (float)NewSampleTime / (float)SampleTime;
         ki *= ratio;
         kd /= ratio;
         SampleTime = (unsigned long)NewSampleTime;
@@ -163,7 +163,7 @@ void PID::SetSampleTime(int NewSampleTime)
  *  want to clamp it from 0-125.  who knows.  at any rate, that can all be done
  *  here.
  **************************************************************************/
-void PID::SetOutputLimits(double Min, double Max)
+void PID::SetOutputLimits(float Min, float Max)
 {
     if (Min >= Max)
         return;
@@ -235,8 +235,8 @@ void PID::SetControllerDirection(int Direction)
  * functions query the internal state of the PID.  they're here for display
  * purposes.  this are the functions the PID Front-end uses for example
  ******************************************************************************/
-double PID::GetKp() { return dispKp; }
-double PID::GetKi() { return dispKi; }
-double PID::GetKd() { return dispKd; }
+float PID::GetKp() { return dispKp; }
+float PID::GetKi() { return dispKi; }
+float PID::GetKd() { return dispKd; }
 int PID::GetMode() { return inAuto ? AUTOMATIC : MANUAL; }
 int PID::GetDirection() { return controllerDirection; }