--- a/pid_v1.h
+++ b/pid_v1.h
@@ -61,7 +61,7 @@ public:
                                       //   means the opposite.  it's very unlikely that this will be needed
                                       //   once it is set in the constructor.
     void SetSampleTime(int);          // * sets the frequency, in Milliseconds, with which
-                                      //   the PID calculation is performed.  default is 100
+                                      //   the PID calculation is performed.  default is 10
 
     //Display functions ****************************************************************
     float GetKp();      // These functions query the pid for interal values.