--- a/sensors_control.cpp
+++ b/sensors_control.cpp
@@ -141,8 +141,6 @@ void updateFunctionSwitch()
     Switch_ADC_value = analogRead(FUNCTION_PIN);
 }
 
-
-
 char emitter_on = 1;
 
 void sensors_control_setup()