@micromouseonline 2021-02-23
The culprit for changes in systick execution times seems to be the Arduino analogWrite function.

After careful timing, pretty much everything run, or close to, constant time until you get right down to that library call.

In the source code, they special-case a zero PWM value into a digital write of a zero to the pin. At all other times, there is additional code.

Not much point in optimising it by replacing the code unless time get really tight. 

Something to bear in mind though and worth a comment in the code.
—
Peter Harrison
peter.harrison@helicron.net



> On 23 Feb 2021, at 10:08, Rob Probin <notifications@github.com> wrote:
> 
> 
> Merged #10 into master.
> 
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub, or unsubscribe.
>