@micromouseonline 2021-02-23
More investigation indicates that the maths libraries may be doing optimisations so that calculations are not constant time. For example, I think a multiply by zero and add zero is optimised. After reset the errors will be zero and several calculations will get short-circuited.

For example, the line

     outputSum += (ki * error);

has two operations that can be skipped if error == 0

Here is the PID compute() method with comment indicating the elapsed time since the beginning of execution for the method


bool PID::Compute()
{
    if (!inAuto)
        return false;
    /*Compute all the working error variables*/
    float input = *myInput;
    float error = *mySetpoint - input;
    float dInput = (input - lastInput);
    // 15.4us -> 17.2us
    outputSum += (ki * error);
    // 26us -> 36us

    /*Add Proportional on Measurement, if P_ON_M is specified*/
    if (!pOnE)
        outputSum -= kp * dInput;

    if (outputSum > outMax)
        outputSum = outMax;
    else if (outputSum < outMin)
        outputSum = outMin;

    // 39us -> 49us
    /*Add Proportional on Error, if P_ON_E is specified*/
    float output;
    if (pOnE)
        output = kp * error;
    else
        output = 0;

    /*Compute Rest of PID Output*/
    output += outputSum - kd * dInput;
    // 61us -> 80us

    if (output > outMax)
        output = outMax;
    else if (output < outMin)
        output = outMin;
    *myOutput = output;

    /*Remember some variables for next time*/
    lastInput = input;
    // 71us -> 90us
    return true;
}
—
Peter Harrison
peter.harrison@helicron.net



> On 23 Feb 2021, at 10:08, Rob Probin <notifications@github.com> wrote:
> 
> 
> Merged #10 into master.
> 
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub, or unsubscribe.
>